requirements.txt
bs4==0.0.1
Django==3.1.4
html5lib==1.1
yfinance==0.1.55
djangorestframework==3.12.2
django-cors-headers==3.6.0
django-filter==2.4.0
graphene==2.1.8
graphene-django==2.14.0
python-jose==3.2.0 # authentication package
